+++ src/main/java/com/nosql/springboot/cassandara/domain/employee/EmployeeRepository.java
@@ -2,5 +2,5 @@ package com.nosql.springboot.cassandara.domain.employee;
 
 import org.springframework.data.cassandra.repository.CassandraRepository;
 
-public interface EmployeeRepository extends CassandraRepository<Employee, Long> {
+public interface EmployeeRepository extends CassandraRepository<Employee, Integer> {
 }
