+++ src/main/java/com/nosql/springboot/cassandara/domain/project/ProjectRepository.java
@@ -2,5 +2,5 @@ package com.nosql.springboot.cassandara.domain.project;
 
 import org.springframework.data.cassandra.repository.CassandraRepository;
 
-public interface ProjectRepository extends CassandraRepository<Project, Long> {
+public interface ProjectRepository extends CassandraRepository<Project, Integer> {
 }
